requirements.txt
streamlit = 1.46.1
tensorflow = 2.19.0
plotly
numpy
Pillowgit